+++ changelog.md
@@ -13,4 +13,16 @@
 - models.py [Model of the City Object]
 - schemas.py [Schema from the City Model]
 - requirements.txt [The tools the API needs to function]
-- helpers.py [Queries for the database]
+- helpers.py [Queries for the database 'SELECT']
+
+## [0.2.5] - 2020-10-10
+
+### Added
+- Blueprints/states.py [Routes for the Table 'estados']
+- Blueprints/messages.py [Routes for the Table 'messages']
+
+### Changes
+- models.py [Models of the State Object, Model of the Message Object]
+- schemas.py [Schema from the State Model, Schema from the Message Model]
+- requirements.txt [Deleted unused libraries]
+- helpers.py [Queries for the database 'INSERT' and Field Verification]